#1071, #1099: Revised Minimal Pairs Search and CSV export.
Moved row code to csv_interface. Revised form buttons in template. Revised show_all setting integration.
